[*] Вынес штрафы и результаты продаж в отдельные роутеры.
