+++ app/api/endpoints/fin_reports.py
@@ -2,9 +2,8 @@ from typing import Optional
 
 from fastapi import APIRouter, Depends, status, Query
 
-from app.dependencies import get_fin_reports_service, get_dates_period_filter, get_months_filter
-from app.domain.models import (WeeklyFinReportsAggregated, DaylyPenaltiesReport,
-                               PeriodRequestModel, MonthlyCategorySales)
+from app.dependencies import get_fin_reports_service, get_dates_period_filter
+from app.domain.models import WeeklyFinReportsAggregated, PeriodRequestModel
 from app.service.fin_reports import FinReportsService
 
 
@@ -21,22 +20,3 @@ async def get_weekly_fin_reports_agg(
     service: FinReportsService = Depends(get_fin_reports_service),
 ) -> list[WeeklyFinReportsAggregated]:
     return await service.get_fin_reports_aggregated(period, number_of_last_weeks)
-
-
-@router.get("/penalties/details", status_code=status.HTTP_200_OK,
-            description="Штрафы по каждому дню")
-async def get_penalties_details(
-    period: PeriodRequestModel = Depends(get_dates_period_filter),
-    service: FinReportsService = Depends(get_fin_reports_service),
-) -> list[DaylyPenaltiesReport]:
-    return await service.get_penalties_details(period)
-
-
-@router.get("/sales_results", status_code=status.HTTP_200_OK,
-            description="Результаты продаж по месяцам и категориям")
-async def get_category_sales_per_month(
-    period: tuple[str] = Depends(get_months_filter),
-    category: Optional[str] = Query(None, description="Категория для фильтрации"),
-    service: FinReportsService = Depends(get_fin_reports_service),
-    ) -> list[MonthlyCategorySales]:
-    return await service.get_category_sales_per_month(*period, category)
