+++ app/dependencies/__init__.py
@@ -8,6 +8,8 @@ from .percent_by_tax import get_percent_by_tax_service, get_percent_by_tax_repos
 from .stocks_quantity import get_stocks_quantity_service, get_stocks_quantity_repository
 from .turnover import get_turnover_service, get_turnover_repository
 from .fin_reports import get_fin_reports_service
+from .penalties import get_penalty_service
+from . sales import get_sale_service
 from .filters import get_dates_period_filter, get_months_filter
 
 
@@ -33,4 +35,6 @@ __all__ = [
     "get_fin_reports_service",
     "get_dates_period_filter",
     "get_months_filter",
+    "get_penalty_service",
+    "get_sale_service",
 ]
