+++ main.py
@@ -7,7 +7,8 @@ import uvicorn
 from app.infrastructure.database import init_db, close_db
 from app.api.endpoints import (article_router, card_data_router, price_discount_router, favicon_router, 
                                turnover_router, orders_revenues_router, unit_economics_router, net_profit_router, 
-                               percent_by_tax_router, stocks_quantity_router, product_router, fin_reports_router)
+                               percent_by_tax_router, stocks_quantity_router, product_router, fin_reports_router,
+                               sales_router, penalties_router)
 from app.config.settings import settings
 
 
@@ -35,6 +36,8 @@ app.include_router(percent_by_tax_router, prefix="/api")
 app.include_router(stocks_quantity_router, prefix="/api")
 app.include_router(product_router, prefix="/api")
 app.include_router(fin_reports_router, prefix="/api")
+app.include_router(penalties_router, prefix="/api")
+app.include_router(sales_router, prefix="/api")
 app.include_router(favicon_router)
 
 origins = [
